feat: add Bugsnag-Integrity header to session payloads
