+++ bugsnag-android-core/src/main/java/com/bugsnag/android/SessionTracker.java
@@ -291,7 +291,7 @@ class SessionTracker extends BaseObservable {
     }
 
     DeliveryStatus deliverSessionPayload(Session payload) {
-        DeliveryParams params = configuration.getSessionApiDeliveryParams();
+        DeliveryParams params = configuration.getSessionApiDeliveryParams(payload);
         Delivery delivery = configuration.getDelivery();
         return delivery.deliver(payload, params);
     }
